+++ src/test/java/racingcar/domain/CarsTest.java
@@ -3,7 +3,7 @@ package racingcar.domain;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import racingcar.util.CarListGenerator;
+import racingcar.utils.CarListGenerator;
 
 import java.util.List;
