+++ src/test/java/racingcar/domain/JudgementTest.java
@@ -2,7 +2,7 @@ package racingcar.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import racingcar.util.CarListGenerator;
+import racingcar.utils.CarListGenerator;
 
 import java.util.List;
